+++ crates/cli/src/commands/pin.rs
@@ -19,7 +19,7 @@ pub struct PinArgs {
     pub global: bool,
 }
 
-pub fn internal_pin(tool: &mut Tool, args: &PinArgs) -> SystemResult {
+pub fn internal_pin(tool: &mut Tool, args: &PinArgs, link: bool) -> SystemResult {
     if args.global {
         tool.manifest.default_version = Some(args.spec.clone());
         tool.manifest.save()?;
@@ -30,8 +30,10 @@ pub fn internal_pin(tool: &mut Tool, args: &PinArgs) -> SystemResult {
             "Wrote the global version",
         );
 
-        // Set symlink to this new version
-        tool.setup_bin_link(true)?;
+        // Create symlink to this new version
+        if link {
+            tool.setup_bin_link(true)?;
+        }
     } else {
         let mut config = ToolsConfig::load()?;
         config.tools.insert(args.id.clone(), args.spec.clone());
@@ -51,7 +53,7 @@ pub fn internal_pin(tool: &mut Tool, args: &PinArgs) -> SystemResult {
 pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
     let mut tool = load_tool(&args.id).await?;
 
-    internal_pin(&mut tool, args)?;
+    internal_pin(&mut tool, args, false)?;
 
     info!(
         "Set the {} version to {}",